requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-crispy-forms==2.1
django-import-export==3.3.3
et-xmlfile==1.1.0
idna==3.4
joblib==1.3.2
MarkupPy==1.14
nltk==3.8.1
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
